Extract `LegacyBufferState`
